Correct markup data handling.
The objects for handling markup data in Video objects are created when
the video entry is pushed to the database.  The metadata import routines
require a database entry to operate against, while the filename
generator requires the import routine to be run before a filename can be
specified.  This in turn requires a dummy entry be created with a blank
filename, which results in a broken markup handler.

This fix manually corrects the markup handler object after a filename
has been chosen, to map it to the correct video.

Fix from CiaranG