Rework metadata selection in mythvidexport
This alteres the metadata selection strategy in mythvidexport.py
following the availability of more complete metadata with the
`mythmetadatalookup` utility. If inetref has been set on the recording,
it will now pull the metadata directly from the recording, rather than
interfacing with the external grabber scripts.  The previously disabled
--listingonly option now works again, forcing this behavior.

WARNING:

The pull of metadata from the recording is performed using the
exportMetadata method from the python bindings, which is currently
incomplete.  Exports of recordings that have been updated using
`mythmetadatalookup` performed using the --listingonly option WILL NOT
WORK until that method has been updated.