Handles issue with 'forgotten' recordings
This adds another type of recording for find_orphans.py to handle.  If a
recording is in the process of deletion when the backend is terminated,
the file and database entry will still exist, but MythTV will ignore any
pre-existing entries with the 'deletepending' flag set. This now deletes
any such recordings over a certain age.